read 순환 제거

+++ TcpServer/Server.cs
@@ -33,7 +33,7 @@ public class Server : TcpServer
 	//listening 전 서버 시작시 필요한 기본 세팅
 	protected override void OnStart() 
 	{
-		PacketHandler.Handler.Add(PacketId.CHAT, ChatCallback);
+		PacketHandler.Handler.Add((ushort)PacketId.CHAT, ChatCallback);
 	}
 	// listen 호출 후
 	protected override void OnStarted() { }
